refactor: move SARIF assembly to output
