+++ crates/output/Cargo.toml
@@ -15,6 +15,7 @@ readme = "../../README.md"
 doctest = false
 
 [dependencies]
+fallow-config = { workspace = true }
 fallow-types = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
